+++ sdk/src/types/block/output/feature/mod.rs
@@ -318,7 +318,7 @@ pub mod dto {
         tag::dto::TagFeatureDto,
     };
     use super::*;
-    use crate::types::block::error::dto::DtoError;
+    use crate::types::block::Error;
 
     #[derive(Clone, Debug, Eq, PartialEq, From)]
     pub enum FeatureDto {
@@ -424,17 +424,17 @@ pub mod dto {
     }
 
     impl TryFrom<&FeatureDto> for Feature {
-        type Error = DtoError;
+        type Error = Error;
 
         fn try_from(value: &FeatureDto) -> Result<Self, Self::Error> {
             Ok(match value {
                 FeatureDto::Sender(v) => Self::Sender(SenderFeature::new((&v.address).try_into()?)),
                 FeatureDto::Issuer(v) => Self::Issuer(IssuerFeature::new((&v.address).try_into()?)),
                 FeatureDto::Metadata(v) => Self::Metadata(MetadataFeature::new(
-                    prefix_hex::decode(&v.data).map_err(|_e| DtoError::InvalidField("MetadataFeature"))?,
+                    prefix_hex::decode(&v.data).map_err(|_e| Error::InvalidField("MetadataFeature"))?,
                 )?),
                 FeatureDto::Tag(v) => Self::Tag(TagFeature::new(
-                    prefix_hex::decode(&v.tag).map_err(|_e| DtoError::InvalidField("TagFeature"))?,
+                    prefix_hex::decode(&v.tag).map_err(|_e| Error::InvalidField("TagFeature"))?,
                 )?),
             })
         }