+++ sdk/src/types/block/signature/mod.rs
@@ -50,7 +50,7 @@ pub mod dto {
 
     pub use super::ed25519::dto::Ed25519SignatureDto;
     use super::*;
-    use crate::types::block::error::dto::DtoError;
+    use crate::types::block::Error;
 
     /// Describes all the different signature types.
     #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, From)]
@@ -68,7 +68,7 @@ pub mod dto {
     }
 
     impl TryFrom<&SignatureDto> for Signature {
-        type Error = DtoError;
+        type Error = Error;
 
         fn try_from(value: &SignatureDto) -> Result<Self, Self::Error> {
             match value {