+++ sdk/CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `NativeTokensBuilder::finish_set`;
 - `Features`, `UnlockConditions`, `NativeTokens`, `MilestoneOptions`, and `Parents` added `from_set`;
+- `types::block::Error::InvalidField` variant;
 
 ### Changed
 
@@ -43,7 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- Remove `AddressGenerationOptions` in favor of `GenerateAddressOptions`, which now contains the `internal` flag;
+- `AddressGenerationOptions` in favor of `GenerateAddressOptions`, which now contains the `internal` flag;
+- `types::block::DtoError`, `client::Error::BlockDto` and `wallet::Error::BlockDto`;
 
 ## 0.2.0 - 2023-04-17